football_shoes_shop/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary
psycopg2
Pillow
django-filter
djangorestframework-simplejwt
stripe
drf-yasg
django-redis
celery
pytest
pytest-django
django-debug-toolbar
python-dotenv
